A first naive trace validation mapping:
The approach by Ron Pressler [1] does not work well for validating a trace that is extremely partial, i.e., it only contains messages that cross the network.  EWD998's trace neither has the values of all variables nor logs every state of the system.  Using non-determinism to define the missing values and states is not practical, as we end up in a TLA+ fragment that TLC cannot handle, such as the following sub-formula of the next-state relation, where the first conjunct defines an infinite set of functions from which the second conjunction filters an infinite subset of functions.

```tla
SomeAct ==
    /\ ...
    /\ inbox' \in [Node -> Seq(Msg)]
    /\ \E idx \in DOMAIN inbox[Initiator]':
            inbox[Initiator][idx]' = Trace[TLCGet("level")].msg
    /\ ...
```

A workaround is to define TLC state and action constraints that match each line in the trace to the current state of a prefix of some behavior defined by the original spec.  The trace can only be mapped to a behavior iff TLC can match every line in the trace to a state in the prefix.  If not, a TLC postcondition will be violated the unmatched logline.

[1] https://www.youtube.com/watch?v=TP3SY0EUV2A