Trace validation found another divergence within ~30 seconds after generating approximately 10k states.
The initiator initiates another token round even when the previous token round is conclusive. The system is in this state after a token round with all nodes white and inactive except for an active but white initiator.

This is a first attempt at a bugfix.

```
<< "Failed matching the trace to (a prefix of) a behavior:",
   [ pkt |->
         [ vc |->
               [ 1 |-> 342,
                 3 |-> 300,
                 0 |-> 216,
                 2 |-> 341,
                 5 |-> 313,
                 4 |-> 313,
                 6 |-> 320,
                 7 |-> 413 ],
           msg |-> [type |-> "pl"],
           snd |-> 0,
           rcv |-> 7 ],
     event |-> "<" ],
   "TLA+ debugger breakpoint hit count 2576" >>  FALSE
Error: Assumption line 212, col 5 to line 215, col 84 of module EWD998ChanTrace is false.
...
```